LayoutTests/media/video-seeking.html

<video controls></video>
<p>Test that seeking attribute is true in timeupdate event after seek and goes back to false when seeking completes </p>
<script src=video-test.js></script>
<script>
waitForEvent('load',
    function () 
    {
        testExpected("video.seeking", false);
        waitForEvent('timeupdate', function () {
            testExpected("video.seeking", true);
            testExpected("video.currentTime", 2);
            setTimeout(function () {
                testExpected("video.seeking", false);
                endTest();
            }, 200);
        });
        run("video.currentTime = 2");
    }
);
video.src = "content/test.mp4";
</script>
