LayoutTests/media/video-seeking-expected.txt

Test that seeking attribute is true in timeupdate event after seek and goes back to false when seeking completes

EVENT(load)
EXPECTED (video.seeking == 'false') OK
RUN(video.currentTime = 2)
EVENT(timeupdate)
EXPECTED (video.seeking == 'true') OK
EXPECTED (video.currentTime == '2') OK
EXPECTED (video.seeking == 'false') OK
END OF TEST
